LPD-53896 create useFDSDrop hook to abstract out common logic. Use it from Table and List visualizations
